requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
asyncpg-lite
beautifulsoup4
lxml
requests
html5lib
fake-headers
psycopg2
